split chatRoomResolver to other file
